Fix all compiler warnings for implicit narrowing conversions
Add explicit casts for int-to-uint8_t/uint16_t conversions in uri.h
and client.cpp. Change input_map::get/set signatures from uint8_t to
int to match call sites and add negative bounds checks.

+++ NetplayInputPlugin/uri.h
@@ -22,7 +22,7 @@ struct uri {
         }
         index = str.rfind(":");
         if (index != std::string::npos && (index == str.find(":") || str[index - 1] == ']')) {
-            port = std::stoi(str.substr(index + 1));
+            port = static_cast<uint16_t>(std::stoi(str.substr(index + 1)));
             str = str.substr(0, index);
         }
         host = str;
